FIFO paths netting zero

+++ amount_status.py
@@ -96,8 +96,8 @@ class status_amounts_inloop:
             self.addrs_src = v[3]
             self.lives_trk = int(v[2])
             self.lives_src = int(v[5])
-            self.lives_trk_updated = float(v[8])
-            self.lives_src_updated = float(v[9])
+            self.lives_trk_updated = int(v[8])
+            self.lives_src_updated = int(v[9])
         else:
             self.status_trk = v[4]
             self.status_src = v[1]
@@ -105,8 +105,8 @@ class status_amounts_inloop:
             self.addrs_src = v[0]
             self.lives_trk = int(v[5])
             self.lives_src = int(v[2])
-            self.lives_trk_updated = float(v[9])
-            self.lives_src_updated = float(v[8])
+            self.lives_trk_updated = int(v[9])
+            self.lives_src_updated = int(v[8])
 
         self.amount_trd = int(v[6])
         self.matched_price = int(v[7])
